requirements.txt
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.5
pydantic>=2.5.0
rich>=13.7.0
sqlite-utils>=3.36
pandas>=2.1.0
tabulate>=0.9.0
python-dotenv>=1.0.0
click>=8.1.0
pyyaml>=6.0.1
matplotlib>=3.8.0
seaborn>=0.13.0
numpy>=1.26.0
